requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.98.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==23.0.0

# Security & Authentication
PyJWT==2.10.1
cryptography==44.0.0
bcrypt==4.2.1
passlib==1.7.4

# Validation & Serialization
marshmallow==3.24.1
cerberus==1.3.5
email-validator==2.3.0

# Database & ORM Enhancements
alembic==1.14.0
SQLAlchemy-Utils==0.42.3

# Logging & Monitoring
loguru==0.7.3
prometheus-client==0.22.0

# Cache & Performance
redis==5.2.1
python-memcached==1.62

# Task Queue & Scheduling
celery==5.4.0
schedule==1.2.2

# Environment & Configuration
python-dotenv==1.0.1
python-decouple==3.9

# HTTP & API Enhancements
requests-toolbelt==1.0.0
httpx[http2]==0.28.1
aiofiles==24.1.0

# File Processing
Pillow==11.0.0
python-magic==0.4.27
PyPDF2==3.1.1

# Testing
pytest==8.4.1
pytest-flask==1.3.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Documentation
flask-restx==1.3.0
flasgger==0.9.7.1

# Utilities
python-slugify==8.0.5
faker==33.1.0
arrow==2.0.1
humanize==4.11.0
